tests/requirements.txt
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
requests-mock>=1.10.0
moto>=4.1.0
pandas>=1.3.0
python-dotenv>=1.0.0
snowflake-connector-python>=3.0.0
boto3>=1.26.0